ENHANCEMENT: Renamed class to Settings, moved it to the Controllers namespace and documented it.
